Simplification of pooltest fsm. Split parts into helper file, und use python math module
